Modified ally.py character.py and enemy.py Move Enemy and Ally class into character.py to avoid circular import problem
